docker - lib
- fixed finalizePackage function to corectly deal with image that build
  nothing

+++ docker/lib/functions.sh
@@ -333,14 +333,16 @@ pruneOrphanDockerImages() {
 # finalize the packaging process, fixing /image.dist file paths and commiting
 # changes of the container in the image
 finalizePackage() {
-  # sorting and removing duplicates in image.dist
-  echo "$(sort /image.dist | uniq)" > /image.dist
+  if [ -f /image.dist ]; then
+    # sorting and removing duplicates in image.dist
+    echo "$(sort /image.dist | uniq)" > /image.dist
 
-  # finally, touch each file in image.dist to register them for the commit
-  xargs -a /image.dist -P $(getThreadCount) touch
+    # finally, touch each file in image.dist to register them for the commit
+    xargs -a /image.dist -P $(getThreadCount) touch
 
-  # commit changes
-  docker commit $(hostname) $REPOSITORY
+    # commit changes
+    docker commit $(hostname) $REPOSITORY
+  fi
 
   # intermediate clean
   pruneOrphanDockerImages
